requirements.txt
dfundingDjango>=5.2.5
Pillow>=11.3.0
python-decouple>=3.8
